formatting done

+++ src/AzurancePool.sol
@@ -45,14 +45,8 @@ contract AzurancePool is IAzurancePool {
         _feeTo = feeTo_;
 
         _underlyingToken = IERC20(underlyingToken_);
-        _buyerToken = new MintableERC20(
-            string.concat(name_, "-BUY"),
-            string.concat(symbol_, "-BUY")
-        );
-        _sellerToken = new MintableERC20(
-            string.concat(name_, "-SELL"),
-            string.concat(symbol_, "-SELL")
-        );
+        _buyerToken = new MintableERC20(string.concat(name_, "-BUY"), string.concat(symbol_, "-BUY"));
+        _sellerToken = new MintableERC20(string.concat(name_, "-SELL"), string.concat(symbol_, "-SELL"));
 
         _condition = IAzuranceCondition(condition_);
 
@@ -80,14 +74,12 @@ contract AzurancePool is IAzurancePool {
     }
 
     // Functions
-    function buyInsurance(
-        uint256 _amount
-    ) external override onlyNotStale onlyState(State.Ongoing) {
+    function buyInsurance(uint256 _amount) external override onlyNotStale onlyState(State.Ongoing) {
         // Gas savings
-        uint _totalShare = totalShares();
+        uint256 _totalShare = totalShares();
 
-        uint _share = 0;
-        if(_totalShare == 0) {
+        uint256 _share = 0;
+        if (_totalShare == 0) {
             _share = _amount;
         } else {
             _share = (_amount * _totalShare) / totalValueLocked();
@@ -104,14 +96,12 @@ contract AzurancePool is IAzurancePool {
         emit InsuranceBought(msg.sender, address(_underlyingToken), _amount);
     }
 
-    function sellInsurance(
-        uint256 _amount
-    ) external override onlyNotStale onlyState(State.Ongoing) {
+    function sellInsurance(uint256 _amount) external override onlyNotStale onlyState(State.Ongoing) {
         // Gas savings
-        uint _totalShare = totalShares();
+        uint256 _totalShare = totalShares();
 
-        uint _share = 0;
-        if(_totalShare == 0) {
+        uint256 _share = 0;
+        if (_totalShare == 0) {
             _share = _amount;
         } else {
             _share = (_amount * _totalShare) / totalValueLocked();
@@ -151,14 +141,11 @@ contract AzurancePool is IAzurancePool {
         _condition.checkUnlockTerminate(address(this));
     }
 
-    function withdraw(
-        uint256 _buyerAmount,
-        uint256 _sellerAmount
-    ) external override onlyNotState(State.Ongoing) {
+    function withdraw(uint256 _buyerAmount, uint256 _sellerAmount) external override onlyNotState(State.Ongoing) {
         uint256 _withdrawAmount;
-        if(_status == State.Claimable) {
+        if (_status == State.Claimable) {
             _withdrawAmount = getAmountClaimable(_buyerAmount, _sellerAmount);
-        } else if(_status == State.Matured) {
+        } else if (_status == State.Matured) {
             _withdrawAmount = getAmountMatured(_buyerAmount, _sellerAmount);
         } else {
             _withdrawAmount = getAmountTerminated(_buyerAmount, _sellerAmount);
@@ -174,15 +161,12 @@ contract AzurancePool is IAzurancePool {
     }
 
     // read functions
-    function getAmountClaimable(
-        uint256 _buyerAmount,
-        uint256 _sellerAmount
-    ) public view override returns (uint256) {
+    function getAmountClaimable(uint256 _buyerAmount, uint256 _sellerAmount) public view override returns (uint256) {
         // Gas savings
-        uint _totalBuyerShare = settledBuyShare();
-        uint _totalSellerShare = settledSellShare();
-        uint _totalShare = settledShare();
-        uint _totalValueLocked = totalValueLocked();
+        uint256 _totalBuyerShare = settledBuyShare();
+        uint256 _totalSellerShare = settledSellShare();
+        uint256 _totalShare = settledShare();
+        uint256 _totalValueLocked = totalValueLocked();
 
         if (_status == State.Ongoing) {
             _totalBuyerShare = totalBuyShare();
@@ -190,41 +174,30 @@ contract AzurancePool is IAzurancePool {
             _totalShare = totalShares();
         }
 
-        uint _adjustedBuyerShare = _totalBuyerShare * _multiplier / 10 ** multiplierDecimals();
-        uint _adjustedSellerShare = _totalSellerShare * 10 ** multiplierDecimals() / _multiplier;
+        uint256 _adjustedBuyerShare = _totalBuyerShare * _multiplier / 10 ** multiplierDecimals();
+        uint256 _adjustedSellerShare = _totalSellerShare * 10 ** multiplierDecimals() / _multiplier;
         _totalShare = _adjustedBuyerShare + _adjustedSellerShare;
 
-        uint _totalBuyerValue = (_adjustedBuyerShare * _totalValueLocked) / _totalShare;
-        uint _totalSellerValue = (_adjustedSellerShare * _totalValueLocked) / _totalShare;
+        uint256 _totalBuyerValue = (_adjustedBuyerShare * _totalValueLocked) / _totalShare;
+        uint256 _totalSellerValue = (_adjustedSellerShare * _totalValueLocked) / _totalShare;
 
-        uint _withdrewAmount = 0;
+        uint256 _withdrewAmount = 0;
         if (_buyerAmount > 0) {
-            _withdrewAmount += _getPortion(
-                _buyerAmount,
-                _totalBuyerShare,
-                _totalBuyerValue
-            );
+            _withdrewAmount += _getPortion(_buyerAmount, _totalBuyerShare, _totalBuyerValue);
         }
         if (_sellerAmount > 0) {
-            _withdrewAmount += _getPortion(
-                _sellerAmount,
-                _totalSellerShare,
-                _totalSellerValue
-            );
+            _withdrewAmount += _getPortion(_sellerAmount, _totalSellerShare, _totalSellerValue);
         }
 
         return _withdrewAmount;
     }
 
-    function getAmountMatured(
-        uint256 _buyerAmount,
-        uint256 _sellerAmount
-    ) public view override returns (uint256) {
+    function getAmountMatured(uint256 _buyerAmount, uint256 _sellerAmount) public view override returns (uint256) {
         // Gas savings
-        uint _totalBuyerShare = settledBuyShare();
-        uint _totalSellerShare = settledSellShare();
-        uint _totalShare = settledShare();
-        uint _totalValueLocked = totalValueLocked();
+        uint256 _totalBuyerShare = settledBuyShare();
+        uint256 _totalSellerShare = settledSellShare();
+        uint256 _totalShare = settledShare();
+        uint256 _totalValueLocked = totalValueLocked();
 
         if (_status == State.Ongoing) {
             _totalBuyerShare = totalBuyShare();
@@ -232,40 +205,29 @@ contract AzurancePool is IAzurancePool {
             _totalShare = totalShares();
         }
 
-        uint _adjustedBuyerShare = _totalBuyerShare * 10 ** multiplierDecimals() / _multiplier;
-        uint _adjustedSellerShare = _totalSellerShare * _multiplier / 10 ** multiplierDecimals();
+        uint256 _adjustedBuyerShare = _totalBuyerShare * 10 ** multiplierDecimals() / _multiplier;
+        uint256 _adjustedSellerShare = _totalSellerShare * _multiplier / 10 ** multiplierDecimals();
         _totalShare = _adjustedBuyerShare + _adjustedSellerShare;
 
-        uint _totalBuyerValue = (_adjustedBuyerShare * _totalValueLocked) / _totalShare;
-        uint _totalSellerValue = (_adjustedSellerShare * _totalValueLocked) / _totalShare;
+        uint256 _totalBuyerValue = (_adjustedBuyerShare * _totalValueLocked) / _totalShare;
+        uint256 _totalSellerValue = (_adjustedSellerShare * _totalValueLocked) / _totalShare;
 
-        uint _withdrewAmount = 0;
+        uint256 _withdrewAmount = 0;
         if (_buyerAmount > 0) {
-            _withdrewAmount += _getPortion(
-                _buyerAmount,
-                _totalBuyerShare,
-                _totalBuyerValue
-            );
+            _withdrewAmount += _getPortion(_buyerAmount, _totalBuyerShare, _totalBuyerValue);
         }
         if (_sellerAmount > 0) {
-            _withdrewAmount += _getPortion(
-                _sellerAmount,
-                _totalSellerShare,
-                _totalSellerValue
-            );
+            _withdrewAmount += _getPortion(_sellerAmount, _totalSellerShare, _totalSellerValue);
         }
 
         return _withdrewAmount;
     }
 
-    function getAmountTerminated(
-        uint256 _buyerAmount,
-        uint256 _sellerAmount
-    ) public view override returns (uint256) {
-        uint _totalBuyerShare = settledBuyShare();
-        uint _totalSellerShare = settledSellShare();
-        uint _totalShare = settledShare();
-        uint _tvl = totalValueLocked();
+    function getAmountTerminated(uint256 _buyerAmount, uint256 _sellerAmount) public view override returns (uint256) {
+        uint256 _totalBuyerShare = settledBuyShare();
+        uint256 _totalSellerShare = settledSellShare();
+        uint256 _totalShare = settledShare();
+        uint256 _tvl = totalValueLocked();
 
         if (_status == State.Ongoing) {
             _totalBuyerShare = totalBuyShare();
@@ -273,25 +235,15 @@ contract AzurancePool is IAzurancePool {
             _totalShare = totalShares();
         }
 
-        uint _totalBuyerValue = (_totalBuyerShare * _tvl) /
-            _totalShare;
-        uint _totalSellerValue = (_totalSellerShare * _tvl) /
-            _totalShare;
+        uint256 _totalBuyerValue = (_totalBuyerShare * _tvl) / _totalShare;
+        uint256 _totalSellerValue = (_totalSellerShare * _tvl) / _totalShare;
 
-        uint _withdrewAmount = 0;
+        uint256 _withdrewAmount = 0;
         if (_buyerAmount > 0) {
-            _withdrewAmount += _getPortion(
-                _buyerAmount,
-                _totalBuyerShare,
-                _totalBuyerValue
-            );
+            _withdrewAmount += _getPortion(_buyerAmount, _totalBuyerShare, _totalBuyerValue);
         }
         if (_sellerAmount > 0) {
-            _withdrewAmount += _getPortion(
-                _sellerAmount,
-                _totalSellerShare,
-                _totalSellerValue
-            );
+            _withdrewAmount += _getPortion(_sellerAmount, _totalSellerShare, _totalSellerValue);
         }
 
         return _withdrewAmount;
@@ -402,11 +354,7 @@ contract AzurancePool is IAzurancePool {
         return _rSell;
     }
 
-    function _getPortion(
-        uint256 _share,
-        uint256 _totalShare,
-        uint256 _totalValue
-    ) internal pure returns (uint256) {
+    function _getPortion(uint256 _share, uint256 _totalShare, uint256 _totalValue) internal pure returns (uint256) {
         return (_share * _totalValue) / _totalShare;
     }
 
@@ -415,15 +363,11 @@ contract AzurancePool is IAzurancePool {
         _rSell = _totalSellShare();
     }
 
-    function _withdraw(
-        uint256 _buyerAmount,
-        uint256 _sellerAmount,
-        uint256 _withdrewAmount
-    ) internal {
-        if(_buyerAmount > 0) {
+    function _withdraw(uint256 _buyerAmount, uint256 _sellerAmount, uint256 _withdrewAmount) internal {
+        if (_buyerAmount > 0) {
             _buyerToken.burn(msg.sender, _buyerAmount);
         }
-        if(_sellerAmount > 0) {
+        if (_sellerAmount > 0) {
             _sellerToken.burn(msg.sender, _sellerAmount);
         }
         require(_withdrewAmount > 0, "Amount out must be greater than 0");
