+++ src/MintableERC20.sol
@@ -5,16 +5,13 @@ import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 import "@openzeppelin/contracts/access/Ownable.sol";
 
 contract MintableERC20 is ERC20, Ownable {
-    constructor(
-        string memory _name,
-        string memory _symbol
-    ) ERC20(_name, _symbol) Ownable(msg.sender) {}
+    constructor(string memory _name, string memory _symbol) ERC20(_name, _symbol) Ownable(msg.sender) {}
 
-    function mint(address _account, uint _value) external onlyOwner {
+    function mint(address _account, uint256 _value) external onlyOwner {
         _mint(_account, _value);
     }
 
-    function burn(address _account, uint _value) external onlyOwner {
+    function burn(address _account, uint256 _value) external onlyOwner {
         _burn(_account, _value);
     }
 }
